chore: moved mock data
